requirements.txt
langchain==0.1.3
openai==1.9.0
pypdf==4.0.0
python-dotenv==1.0.1
streamlit==1.30.0
spacy==3.7.2